requirements.txt
anthropic
anyio
boto3
certifi
click
colorama
diskcache
diskcache-stubs
distro
exceptiongroup
execnet
falkordb
fastapi
fastapi-cli
google-genai
graphiti-core
groq
h11
hatchling
httpcore
httptools
httpx
ipykernel
jupyterlab
kuzu
langchain-anthropic
langchain-aws
langchain-openai
langgraph
langsmith
neo4j
neptune
numpy
openai
opensearch-py
opentelemetry-api
opentelemetry-sdk
posthog
pydantic
pyright
pytest
pytest-asyncio
pytest-xdist
python-dotenv
ruff
sentence-transformers
tenacity
tracing
transformers
voyageai
